+++ Leviasan.Sanlog.Abstractions/FormattedLogValuesFormatter.cs
@@ -10,9 +10,12 @@ using System.Text;
 namespace Leviasan.Sanlog
 {
     /// <summary>
-    /// Represents the formatter of the <see cref="Microsoft.Extensions.Logging.FormattedLogValues"/> class.
-    /// Provides a mechanism to parse named format string to composite string.
+    /// Represents the formatter of the Microsoft.Extensions.Logging.FormattedLogValues class.
     /// </summary>
+    /// <remarks>
+    /// Overrides a default format string for <see cref="DateTime"/> to a round-trip date/time pattern defined in ISO 8601 ("O") and for <see cref="Enum"/> to the shortest integer value representation possible ("D").
+    /// Overrides a string representation of the <see cref="IDictionary"/> and <see cref="IEnumerable"/> objects.
+    /// </remarks>
     public sealed class FormattedLogValuesFormatter : IReadOnlyList<KeyValuePair<string, string>>
     {
         /// <summary>
@@ -31,65 +34,76 @@ namespace Leviasan.Sanlog
         /// The message format that represents a structured logging message.
         /// </summary>
         public const string OriginalFormat = "{OriginalFormat}";
-        /// <summary>
-        /// The delimiters used by NamedFormat composite string.
-        /// </summary>
-        [DebuggerBrowsable(DebuggerBrowsableState.Never)]
-        private static readonly char[] FormatDelimiters = [',', ':'];
 
         /// <summary>
         /// Max cached collection size.
         /// </summary>
         [DebuggerBrowsable(DebuggerBrowsableState.Never)]
-        private const int MaxCachedCollectionSize = 1024;
+        private const int MaxCachedFormatters = 1024; // Microsoft.Extensions.Logging.FormattedLogValues.MaxCachedFormatters
         /// <summary>
-        /// The cache collection of the composite strings.
+        /// The cache of the named format strings.
         /// </summary>
         [DebuggerBrowsable(DebuggerBrowsableState.Never)]
-        private static readonly Dictionary<string, CompositeFormat> CachedCompositeFormat = [];
+        private static readonly Dictionary<string, NamedFormatString> CachedNamedFormatStrings = [];
         /// <summary>
-        /// The cache collection of the value names.
+        /// Tries to get a named format string from the cache or parses and tries to add to cache it.
         /// </summary>
-        [DebuggerBrowsable(DebuggerBrowsableState.Never)]
-        private static readonly Dictionary<string, IReadOnlyList<string>> CachedValueNames = [];
+        /// <param name="format">The key of the element to get or add.</param>
+        /// <param name="namedFormatString">When this method returns, it contains a parsed named format string or the <see langword="null"/> if the operation failed.</param>
+        /// <returns><see langword="true"/> if operation is successful; otherwise <see langword="false"/>.</returns>
+        /// <exception cref="FormatException">A format item in format is invalid.</exception>
+        private static bool TryGetOrAdd(string? format, [NotNullWhen(true)] out NamedFormatString? namedFormatString)
+        {
+            if (string.IsNullOrEmpty(format))
+            {
+                namedFormatString = default;
+                return false;
+            }
+            if (CachedNamedFormatStrings.TryGetValue(format, out namedFormatString)) return true;
+
+            namedFormatString = NamedFormatString.Parse(format); // FormatException
+            return namedFormatString.CompositeFormat.MinimumArgumentCount <= 0 || CachedNamedFormatStrings.Count >= MaxCachedFormatters || CachedNamedFormatStrings.TryAdd(format, namedFormatString);
+        }
 
         /// <summary>
-        /// The original property values.
+        /// The hash set of the sensitive segment's names.
         /// </summary>
         [DebuggerBrowsable(DebuggerBrowsableState.Never)]
-        private readonly IReadOnlyList<KeyValuePair<string, object?>> _logValues;
+        private readonly HashSet<string> _sensitiveData = [];
         /// <summary>
-        /// The name collection of the named item format properties that belong to sensitive data.
+        /// The original values.
         /// </summary>
         [DebuggerBrowsable(DebuggerBrowsableState.Never)]
-        private readonly HashSet<string> _sensitiveDataType = [];
-        /// <summary>
-        /// The function to get a string representation of the object.
-        /// </summary>
-        [DebuggerBrowsable(DebuggerBrowsableState.Never)]
-        private readonly Func<string?>? _formatter;
+        private readonly IReadOnlyList<KeyValuePair<string, object?>> _originalValues;
         /// <summary>
         /// An object that supplies culture-specific formatting information.
         /// </summary>
         [DebuggerBrowsable(DebuggerBrowsableState.Never)]
         private readonly IFormatProvider? _formatProvider;
+        /// <summary>
+        /// The parsed named format string.
+        /// </summary>
+        [DebuggerBrowsable(DebuggerBrowsableState.Never)]
+        private readonly NamedFormatString? _namedFormatString;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="FormattedLogValuesFormatter"/> class with the original property values and name collection of the named item format properties that belong to sensitive data.
+        /// Initializes a new instance of the <see cref="FormattedLogValuesFormatter"/> class with an array of original values and an object that supplies culture-specific formatting information.
         /// </summary>
-        /// <param name="formatProvider">An object that supplies culture-specific formatting information. If the passed value is <see langword="null"/> it's the same as using <see cref="CultureInfo.CurrentCulture"/>.</param>
-        /// <param name="original">The original property values.</param>
-        /// <param name="formatter">The function to get string representation of the object if <paramref name="original"/> does not contains key <see cref="OriginalFormat"/>.</param>
-        public FormattedLogValuesFormatter(IFormatProvider? formatProvider, IReadOnlyList<KeyValuePair<string, object?>>? original, Func<string?>? formatter)
+        /// <param name="original">An array of original values.</param>
+        /// <param name="formatProvider">An object that supplies culture-specific formatting information.</param>
+        /// <exception cref="FormatException">A format item in format is invalid.</exception>
+        /// <exception cref="InvalidOperationException">More than one <see cref="OriginalFormat"/> key was found at <paramref name="original"/> array.</exception>
+        public FormattedLogValuesFormatter(IReadOnlyList<KeyValuePair<string, object?>>? original, IFormatProvider? formatProvider)
         {
+            _originalValues = original ?? [];
             _formatProvider = formatProvider ?? CultureInfo.CurrentCulture;
-            _logValues = original ?? [];
-            _formatter = formatter;
+            _ = TryGetOrAdd(_originalValues.SingleOrDefault(x => x.Key == OriginalFormat).Value as string, out _namedFormatString); // InvalidOperationException + FormatException
         }
         /// <summary>
         /// Initializes a new instance of the <see cref="FormattedLogValuesFormatter"/> class with the specified named format string to parse and an object array that contains zero or more objects to format.
         /// </summary>
-        /// <param name="formatProvider">An object that supplies culture-specific formatting information. If the passed value is <see langword="null"/> it's the same as using <see cref="CultureInfo.CurrentCulture"/>.</param>
+        /// <param name="cache">An object that supplies culture-specific formatting information.</param>
+        /// <param name="formatProvider">An object that supplies culture-specific formatting information.</param>
         /// <param name="format">The named format string to parse.</param>
         /// <param name="args">An object array that contains zero or more objects to format.</param>
         /// <exception cref="ArgumentNullException">The <paramref name="format"/> or <paramref name="args"/> is <see langword="null"/>.</exception>
@@ -99,70 +113,122 @@ namespace Leviasan.Sanlog
             ArgumentNullException.ThrowIfNull(format);
             ArgumentNullException.ThrowIfNull(args);
 
-            TryGetValueOrParseNamedFormat(format, out _, out var valueNames);
-            var original = new List<KeyValuePair<string, object?>>();
-            for (var index = 0; index < args.Length; ++index)
-                original.Add(new KeyValuePair<string, object?>(valueNames[index], args[index]));
-            original.Add(new KeyValuePair<string, object?>(OriginalFormat, format));
-
-            _logValues = original;
             _formatProvider = formatProvider ?? CultureInfo.CurrentCulture;
+            if (TryGetOrAdd(format, out _namedFormatString)) // FormatException
+            {
+                var original = new List<KeyValuePair<string, object?>>();
+                for (int index = 0, segmentId = 0; index < args.Length; ++index, ++segmentId)
+                {
+                    while (segmentId < _namedFormatString.Segments.Count && original.Any(x => x.Key == _namedFormatString.Segments[segmentId].Name))
+                    {
+                        ++segmentId;
+                    }
+                    original.Add(KeyValuePair.Create(_namedFormatString.Segments[segmentId].Name, args[index]));
+                }
+                original.Add(KeyValuePair.Create<string, object?>(OriginalFormat, format));
+                _originalValues = original;
+            }
+            else
+            {
+                _originalValues = args.Select((element, index) => KeyValuePair.Create(index.ToString(null, null), element)).ToList();
+            }
         }
 
         /// <inheritdoc/>
         /// <exception cref="ArgumentOutOfRangeException">Index was outside the bounds of the array.</exception>
-        public KeyValuePair<string, string> this[int index] => GetData(index, true);
+        public KeyValuePair<string, string> this[int index] => GetObjectAsString(index, true);
         /// <summary>
-        /// Gets the element value at the specified key in the read-only list.
+        /// Gets the string representation of the element value at the specified key.
         /// </summary>
-        /// <param name="key">The key to find.</param>
+        /// <param name="name">The property name of the element to get.</param>
         /// <returns>The element value at the specified key in the read-only list.</returns>
-        /// <exception cref="ArgumentNullException">The <paramref name="key"/> is <see langword="null"/>.</exception>
-        /// <exception cref="KeyNotFoundException">The <paramref name="key"/> not found.</exception>
-        public string this[string key] => GetData(key, true).Value;
+        /// <exception cref="ArgumentNullException">The <paramref name="name"/> is <see langword="null"/>.</exception>
+        /// <exception cref="KeyNotFoundException">The <paramref name="name"/> not found.</exception>
+        public KeyValuePair<string, string> this[string name] => GetObjectAsString(name, true);
         /// <inheritdoc/>
-        public int Count => _logValues.Count;
+        public int Count => _originalValues.Count;
+        /// <summary>
+        /// Indicates whether <see cref="OriginalFormat"/> is registered.
+        /// </summary>
+        public bool HasOriginalFormat => _originalValues.Any(x => x.Key == OriginalFormat);
 
         /// <inheritdoc/>
         public IEnumerator<KeyValuePair<string, string>> GetEnumerator()
         {
-            for (var index = 0; index < Count; ++index) yield return this[index];
+            for (var index = 0; index < Count; ++index)
+                yield return GetObjectAsString(index, true);
         }
         /// <inheritdoc/>
         IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
         /// <summary>
-        /// Gets the key-value pair at the specified index.
+        /// Gets a key-value pair describing a property name and object.
         /// </summary>
         /// <param name="index">The zero-based index of the element to get.</param>
-        /// <param name="redacted">Indicates whether need redact value after extracting.</param>
-        /// <returns>The key-value pair at the specified index.</returns>
+        /// <param name="redacted">Indicates whether need redact value.</param>
+        /// <returns>The key-value pair describing a property name and object.</returns>
         /// <exception cref="ArgumentOutOfRangeException">Index was outside the bounds of the array.</exception>
-        public KeyValuePair<string, string> GetData(int index, bool redacted)
+        public KeyValuePair<string, object?> GetObject(int index, bool redacted)
         {
-            return new KeyValuePair<string, string>(_logValues[index].Key, redacted && IsSensitiveData(_logValues[index].Key) ? RedactedValue : FormatArgument(_formatProvider, _logValues[index].Value));
+            var key = _originalValues[index].Key;
+            return KeyValuePair.Create(key, redacted && IsSensitiveData(key) ? RedactedValue : _originalValues[index].Value);
+        }
+        /// <summary>
+        /// Gets a key-value pair describing a property name and object.
+        /// </summary>
+        /// <param name="name">The property name to find.</param>
+        /// <param name="redacted">Indicates whether need redact value.</param>
+        /// <returns>The key-value pair describing a property name and object.</returns>
+        /// <exception cref="ArgumentNullException">The <paramref name="name"/> is <see langword="null"/>.</exception>
+        /// <exception cref="KeyNotFoundException">The <paramref name="name"/> is not found.</exception>
+        public KeyValuePair<string, object?> GetObject(string name, bool redacted)
+        {
+            ArgumentNullException.ThrowIfNull(name);
+            for (var index = 0; index < _originalValues.Count; ++index)
+                if (_originalValues[index].Key == name) return GetObject(index, redacted);
+            throw new KeyNotFoundException();
+        }
+        /// <summary>
+        /// Gets a key-value pair describing a property name and string representation of the object.
+        /// </summary>
+        /// <param name="index">The zero-based index of the element to get.</param>
+        /// <param name="redacted">Indicates whether need redact value.</param>
+        /// <returns>The key-value pair describing a property name and string representation of the object.</returns>
+        /// <exception cref="ArgumentOutOfRangeException">Index was outside the bounds of the array.</exception>
+        public KeyValuePair<string, string> GetObjectAsString(int index, bool redacted)
+        {
+            var pair = GetObject(index, redacted);
+            var formattedValue = FormatArgument(_formatProvider, pair.Value);
+            return KeyValuePair.Create(pair.Key, formattedValue);
 
-            // Summary: Formats the specified value to a string representation
-            // Param (formatProvider): An object that supplies culture-specific formatting information
-            // Param (value): An object to format
-            // Returns: A string representation of the formatted object
-            static string FormatArgument(IFormatProvider? formatProvider, object? value) => TryFormatArgument(formatProvider, value, out var stringValue) ? stringValue : Convert.ToString(value ?? NullValue, formatProvider)!;
-            // Summary: Tries to format the specified value to a string representation
-            // Param (formatProvider): An object that supplies culture-specific formatting information
-            // Param (value): An object to format
-            // Param out (stringValue): A string representation of the formatted object if the operation is successful; otherwise null
-            // Returns: true if the format operation is successful; otherwise false
-            static bool TryFormatArgument<T>(IFormatProvider? formatProvider, T? value, [NotNullWhen(true)] out string? stringValue)
+            // Summary: Formats the specified value to a string representation.
+            // Param (formatProvider): An object that supplies culture-specific formatting information.
+            // Param (value): An object to format.
+            // Returns: A string representation of the formatted object.
+            static string FormatArgument(IFormatProvider? formatProvider, object? value)
+                => TryFormatArgument(formatProvider, value, out var stringValue) ? stringValue : Convert.ToString(value ?? NullValue, formatProvider)!;
+            // Summary: Tries to format the specified value to a string representation.
+            // Param (formatProvider): An object that supplies culture-specific formatting information.
+            // Param (value): An object to format.
+            // Param out (stringValue): A string representation of the formatted object if the operation is successful; otherwise null.
+            // Returns: true if the format operation is successful; otherwise false.
+            static bool TryFormatArgument(IFormatProvider? formatProvider, object? value, [NotNullWhen(true)] out string? stringValue)
             {
-                // If the value is DateTime using a sortable date/time pattern ("s") defined in ISO 8601
-                if (value is DateTime dateTime)
+                // If the value is Enum display the enumeration entry as an integer value in the shortest representation possible
+                if (value is Enum enumeration)
                 {
-                    stringValue = dateTime.ToString("s", formatProvider);
+                    stringValue = enumeration.ToString("D");
                     return true;
                 }
-                // If the value is DateTimeOffset using a sortable date/time pattern ("s") defined in ISO 8601
+                // If the value is DateTime using a round-trip date/time pattern defined in ISO 8601
+                else if (value is DateTime dateTime)
+                {
+                    stringValue = dateTime.ToString("O", formatProvider);
+                    return true;
+                }
+                // If the value is DateTimeOffset using a round-trip date/time pattern defined in ISO 8601
                 else if (value is DateTimeOffset dateTimeOffset)
                 {
-                    stringValue = dateTimeOffset.ToString("s", formatProvider);
+                    stringValue = dateTimeOffset.ToString("O", formatProvider);
                     return true;
                 }
                 // If the value implements IDictionary builds a comma-separated string in KeyValuePair->ToString style
@@ -173,7 +239,7 @@ namespace Leviasan.Sanlog
                     foreach (DictionaryEntry entry in dictionary)
                     {
                         if (!first) _ = stringBuilder.Append(", ");
-                        _ = stringBuilder.Append(formatProvider, $"[{Convert.ToString(entry.Key, formatProvider)}, {Convert.ToString(entry.Value ?? NullValue, formatProvider)}]");
+                        _ = stringBuilder.Append(formatProvider, $"[{FormatArgument(formatProvider, entry.Key)}, {FormatArgument(formatProvider, entry.Value)}]");
                         first = false;
                     }
                     stringValue = stringBuilder.ToString();
@@ -187,7 +253,7 @@ namespace Leviasan.Sanlog
                     foreach (var e in enumerable)
                     {
                         if (!first) _ = stringBuilder.Append(", ");
-                        _ = stringBuilder.Append(Convert.ToString(e ?? NullValue, formatProvider));
+                        _ = stringBuilder.Append(FormatArgument(formatProvider, e));
                         first = false;
                     }
                     stringValue = stringBuilder.ToString();
@@ -198,187 +264,75 @@ namespace Leviasan.Sanlog
             }
         }
         /// <summary>
-        /// Gets the key-value pair at the specified key.
+        /// Gets a key-value pair describing a property name and string representation of the object.
         /// </summary>
-        /// <param name="key">The key to find.</param>
-        /// <param name="redacted">Indicates whether need redact value after extracting.</param>
-        /// <returns>The key-value pair at the specified key.</returns>
-        /// <exception cref="ArgumentNullException">The <paramref name="key"/> is <see langword="null"/>.</exception>
-        /// <exception cref="KeyNotFoundException">The <paramref name="key"/> not found.</exception>
-        public KeyValuePair<string, string> GetData(string key, bool redacted)
+        /// <param name="name">The property name to find.</param>
+        /// <param name="redacted">Indicates whether need redact value.</param>
+        /// <returns>The key-value pair describing a property name and string representation of the object.</returns>
+        /// <exception cref="ArgumentNullException">The <paramref name="name"/> is <see langword="null"/>.</exception>
+        /// <exception cref="KeyNotFoundException">The <paramref name="name"/> is not found.</exception>
+        public KeyValuePair<string, string> GetObjectAsString(string name, bool redacted)
         {
-            ArgumentNullException.ThrowIfNull(key);
-            for (var index = 0; index < _logValues.Count; ++index)
-                if (_logValues[index].Key == key) return GetData(index, redacted);
+            ArgumentNullException.ThrowIfNull(name);
+            for (var index = 0; index < _originalValues.Count; ++index)
+                if (_originalValues[index].Key == name) return GetObjectAsString(index, redacted);
             throw new KeyNotFoundException();
         }
         /// <summary>
-        /// Checks whether a value of the assotiated key of the <see cref="KeyValuePair{TKey, TValue}"/> belongs to sensitive data.
+        /// Checks whether a segment name belongs to sensitive data.
         /// </summary>
-        /// <param name="key">The key of the <see cref="KeyValuePair{TKey, TValue}"/>.</param>
-        /// <returns><see langword="true"/> if the value of the property belongs to sensitive data; otherwise <see langword="false"/>.</returns>
-        public bool IsSensitiveData(string key) => _sensitiveDataType.Contains(key);
+        /// <param name="name">The name of the segment to check.</param>
+        /// <returns><see langword="true"/> if the segment name belongs to sensitive data; otherwise <see langword="false"/>.</returns>
+        public bool IsSensitiveData(string name) => _sensitiveData.Contains(name);
         /// <summary>
-        /// Registers a name of the named item format property whose associated value will be redacted before logging.
+        /// Registers a segment name whose associated value will be redacted before logging.
         /// </summary>
-        /// <param name="key">The name of the named item format property that belongs to sensitive data.</param>
+        /// <param name="name">The name of the segment that belongs to sensitive data.</param>
         /// <returns><see langword="true"/> if the element is added to the collection; <see langword="false"/> if the element is already present.</returns>
-        /// <exception cref="ArgumentNullException">The <paramref name="key"/> is <see langword="null"/>.</exception>
-        public bool RegisterSensitiveData(string key)
+        /// <exception cref="ArgumentNullException">The <paramref name="name"/> is <see langword="null"/>.</exception>
+        public bool RegisterSensitiveData(string name)
         {
-            ArgumentNullException.ThrowIfNull(key);
-            return key != OriginalFormat && _sensitiveDataType.Add(key);
+            ArgumentNullException.ThrowIfNull(name);
+            return name != OriginalFormat && _sensitiveData.Add(name);
         }
         /// <summary>
-        /// Registers name collection of the named item format properties whose associated values will be redacted before logging.
+        /// Registers segment names whose associated value will be redacted before logging.
         /// </summary>
-        /// <param name="keys">The name collection of the named item format properties that belong to sensitive data.</param>
-        /// <exception cref="ArgumentNullException">The <paramref name="keys"/> or one of the elements in array is <see langword="null"/>.</exception>
-        public void RegisterSensitiveData(IEnumerable<string> keys)
+        /// <param name="names">The names of the segment that belongs to sensitive data.</param>
+        /// <returns>The count of the added element.</returns>
+        /// <exception cref="ArgumentNullException">The <paramref name="names"/> or at least one element in the specified array is <see langword="null"/>.</exception>
+        public int RegisterSensitiveData(params string[] names)
         {
-            ArgumentNullException.ThrowIfNull(keys);
-            foreach (var key in keys)
-            {
-                if (key is null) throw new ArgumentNullException(nameof(keys));
-                _ = RegisterSensitiveData(key);
-            }
+            ArgumentNullException.ThrowIfNull(names);
+            if (!Array.TrueForAll(names, x => x is not null))
+                throw new ArgumentNullException(nameof(names), "At least one element in the specified array was null.");
+
+            var count = 0;
+            foreach (var name in names)
+                count += Convert.ToInt32(RegisterSensitiveData(name));
+            return count;
         }
-        /// <summary>
-        /// Returns a string representation of the current instance.
-        /// </summary>
-        /// <returns>A string representation of the current instance.</returns>
-        /// <exception cref="FormatException">A format item in format is invalid.</exception>
-        public override string? ToString()
+        /// <inheritdoc/>
+        public override string ToString()
         {
-            if (_logValues.SingleOrDefault(x => x.Key == OriginalFormat).Value is string originalFormat)
-            {
-                TryGetValueOrParseNamedFormat(originalFormat, out var compositeFormat, out var valueNames);
-                var stringValue = compositeFormat.MinimumArgumentCount <= _logValues.Count
-                    ? string.Format(_formatProvider, compositeFormat, TakeByValueNamesOrder(valueNames))
-                    : compositeFormat.Format;
-                return stringValue;
-            }
-            return _formatter is not null ? _formatter.Invoke() : NullFormat;
+            return _namedFormatString is not null
+                ? _namedFormatString.Format(_formatProvider, TakeBySegmentOrder(_namedFormatString.Segments))
+                : NullFormat;
 
-            // Summary: Returns a sequence of redacted values based on an order of value names
-            // Param (valueNames): A sequence of the names in the expression
-            // Returns: A sequence of the redacted values order by value names
-            ReadOnlySpan<object?> TakeByValueNamesOrder(IReadOnlyList<string> valueNames)
+            // Summary: Returns a sequence of redacted values based on an order of segment names.
+            // Param (segments): A sequence of the segments in the named format string.
+            // Returns: A sequence of the redacted values order by segment names.
+            object?[] TakeBySegmentOrder(IReadOnlyList<NamedFormatString.FormatSegment> segments)
             {
-                var array = new string[valueNames.Count];
-                for (var index = 0; index < array.Length; ++index)
-                    array[index] = this[valueNames[index]];
-                return array;
-            }
-        }
-        /// <summary>
-        /// Parses the specified named format string and tries to save it to the cache.
-        /// </summary>
-        /// <param name="format">The named format string to parse.</param>
-        /// <param name="compositeFormat">The parsed composite string.</param>
-        /// <param name="valueNames">The value names collection.</param>
-        /// <exception cref="FormatException">A format item in format is invalid.</exception>
-        private static void TryGetValueOrParseNamedFormat(string format, out CompositeFormat compositeFormat, out IReadOnlyList<string> valueNames)
-        {
-            Debug.Assert(format is not null);
-            if (!CachedCompositeFormat.TryGetValue(format, out compositeFormat!))
-            {
-                (compositeFormat, valueNames) = ParseNamedFormat(format);
-                if (compositeFormat.MinimumArgumentCount > 0 && CachedCompositeFormat.Count <= MaxCachedCollectionSize)
+                var dictionary = new Dictionary<string, object?>();
+                foreach (var segment in segments)
                 {
-                    CachedCompositeFormat.Add(format, compositeFormat);
-                    CachedValueNames.Add(format, valueNames);
+                    if (dictionary.ContainsKey(segment.Name)) continue;
+                    dictionary[segment.Name] = segments.Where(x => x.Name == segment.Name).All(x => string.IsNullOrEmpty(x.FormatString))
+                        ? GetObjectAsString(segment.Name, true).Value
+                        : GetObject(segment.Name, true).Value;
                 }
-            }
-            else
-            {
-                valueNames = CachedValueNames[format];
-            }
-
-            // Summary: Parses the specified named format string to composite format string.
-            // Exception: System.FormatException - A format item in format is invalid.
-            static (CompositeFormat CompositeFormat, IReadOnlyList<string> ValueNames) ParseNamedFormat(string format)
-            {
-                var scanIndex = 0;
-                var endIndex = format.Length;
-                var stringBuilder = new StringBuilder(256);
-                var valueNames = new List<string>();
-                while (scanIndex < endIndex)
-                {
-                    var openBraceIndex = FindBraceIndex(format, '{', scanIndex, endIndex);
-                    if (scanIndex == 0 && openBraceIndex == endIndex)
-                    {
-                        return (CompositeFormat.Parse(format), valueNames);
-                    }
-                    var closeBraceIndex = FindBraceIndex(format, '}', openBraceIndex, endIndex);
-                    if (closeBraceIndex == endIndex)
-                    {
-                        _ = stringBuilder.Append(format.AsSpan(scanIndex, endIndex - scanIndex));
-                        scanIndex = endIndex;
-                    }
-                    else
-                    {
-                        // Format item syntax : { index[,alignment][ :formatString] }.
-                        var formatDelimiterIndex = FindIndexOfAny(format, FormatDelimiters, openBraceIndex, closeBraceIndex);
-                        _ = stringBuilder.Append(format.AsSpan(scanIndex, openBraceIndex - scanIndex + 1));
-                        _ = stringBuilder.Append(valueNames.Count);
-                        _ = stringBuilder.Append(format.AsSpan(formatDelimiterIndex, closeBraceIndex - formatDelimiterIndex + 1));
-                        scanIndex = closeBraceIndex + 1;
-                        valueNames.Add(format.Substring(openBraceIndex + 1, formatDelimiterIndex - openBraceIndex - 1));
-                    }
-                }
-                return (CompositeFormat.Parse(stringBuilder.ToString()), valueNames);
-
-                // Summary: Reports the zero-based index of the first occurrence in string instance of brace in a specified array of Unicode characters.
-                // The search starts at a specified character position and examines a specified number of character positions.
-                static int FindBraceIndex(string format, char brace, int startIndex, int endIndex)
-                {
-                    // Example: {{prefix{{{Argument}}}suffix}}.
-                    var braceIndex = endIndex;
-                    var scanIndex = startIndex;
-                    var braceOccurrenceCount = 0;
-                    while (scanIndex < endIndex)
-                    {
-                        if (braceOccurrenceCount > 0 && format[scanIndex] != brace)
-                        {
-                            if (braceOccurrenceCount % 2 == 0)
-                            {
-                                // Even number of '{' or '}' found. Proceed search with next occurrence of '{' or '}'.
-                                braceOccurrenceCount = 0;
-                                braceIndex = endIndex;
-                            }
-                            else
-                            {
-                                // An unescaped '{' or '}' found.
-                                break;
-                            }
-                        }
-                        else if (format[scanIndex] == brace)
-                        {
-                            if (brace == '}')
-                            {
-                                if (braceOccurrenceCount == 0)
-                                    // For '}' pick the first occurrence.
-                                    braceIndex = scanIndex;
-                            }
-                            else
-                            {
-                                // For '{' pick the last occurrence.
-                                braceIndex = scanIndex;
-                            }
-                            braceOccurrenceCount++;
-                        }
-                        scanIndex++;
-                    }
-                    return braceIndex;
-                }
-                // Summary: Reports the zero-based index of the first occurrence in string instance of any character in a specified array of Unicode characters.
-                // The search starts at a specified character position and examines a specified number of character positions.
-                static int FindIndexOfAny(string format, char[] chars, int startIndex, int endIndex)
-                {
-                    var findIndex = format.IndexOfAny(chars, startIndex, endIndex - startIndex);
-                    return findIndex == -1 ? endIndex : findIndex;
-                }
+                return [.. dictionary.Values];
             }
         }
     }