+++ Leviasan.Sanlog.EntityFrameworkCore/CHANGELOG.md
@@ -8,3 +8,7 @@
 ### 1.1.1
 - `LoggingEntryProperty.Id`, `LoggingScope.Id` and `LoggingScopeProperty.Id` never have a value generated when an instance of this entity type is saved.
 - Updated NuGet dependency `Microsoft.EntityFrameworkCore` from 8.0.4 to 8.0.5.
+
+### 1.1.2
+- MSBuild property `GenerateDocumentationFile` is disabled.
+- Updated NuGet dependency `Microsoft.EntityFrameworkCore` from 8.0.5 to 8.0.6.